requirement.txt
aiohttp==3.6.2
aiomysql==0.0.20
asgiref==3.2.3
astroid==2.2.5
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
beautifulsoup4==4.6.0
certifi==2019.6.16
cffi==1.13.2
chardet==3.0.4
Click==7.0
clickclick==1.2.2
colorama==0.4.1
comtypes==1.1.7
connexion==2.4.0
cryptography==2.8
cursor==1.3.4
Django==3.0.2
django-filter==2.2.0
django-mysql==3.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
Faker==2.0.4
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JSON==0.3.4
future==0.18.2
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
jsonfield==2.0.2
jsonschema==3.1.1
lazy-object-proxy==1.4.1
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mimesis==3.3.0
more-itertools==7.2.0
multidict==4.7.3
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.11
mysqlclient==1.4.6
numpy==1.17.0
oauthlib==3.1.0
openapi-spec-validator==0.2.8
opencv-python==3.4.2.17
openpyxl==2.4.8
passlib==1.7.2
Pillow==7.0.0
protobuf==3.11.1
py-trello==0.13.0
PyAutoGUI==0.9.36
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pymemcache==2.2.2
PyMsgBox==1.0.7
PyMySQL==0.9.2
PyPDF2==1.26.0
pyrsistent==0.15.5
PyScreeze==0.1.22
pytesseract==0.2.0
python-dateutil==2.8.0
python-docx==0.8.6
python-fity3==0.7
python3-memcached==1.51
PyTweening==1.0.3
pytz==2019.3
pywin32==223
pywinauto==0.6.5
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
retrying==1.3.3
selenium==3.7.0
six==1.12.0
socketIO-client==0.7.2
sqlparse==0.3.0
sty==1.0.0b2
swagger-ui-bundle==0.0.6
text-unidecode==1.3
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
websocket-client==0.56.0
Werkzeug==0.16.0
wrapt==1.11.2
XlsxWriter==1.2.6
yarl==1.4.2
zipp==0.6.0
aiohttp==3.6.2
aiomysql==0.0.20
asgiref==3.2.3
astroid==2.2.5
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
beautifulsoup4==4.6.0
certifi==2019.6.16
cffi==1.13.2
chardet==3.0.4
Click==7.0
clickclick==1.2.2
colorama==0.4.1
comtypes==1.1.7
connexion==2.4.0
cryptography==2.8
cursor==1.3.4
Django==3.0.2
django-filter==2.2.0
django-mysql==3.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
Faker==2.0.4
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JSON==0.3.4
future==0.18.2
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
jsonfield==2.0.2
jsonschema==3.1.1
lazy-object-proxy==1.4.1
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mimesis==3.3.0
more-itertools==7.2.0
multidict==4.7.3
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.11
mysqlclient==1.4.6
numpy==1.17.0
oauthlib==3.1.0
openapi-spec-validator==0.2.8
opencv-python==3.4.2.17
openpyxl==2.4.8
passlib==1.7.2
Pillow==7.0.0
protobuf==3.11.1
py-trello==0.13.0
PyAutoGUI==0.9.36
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pymemcache==2.2.2
PyMsgBox==1.0.7
PyMySQL==0.9.2
PyPDF2==1.26.0
pyrsistent==0.15.5
PyScreeze==0.1.22
pytesseract==0.2.0
python-dateutil==2.8.0
python-docx==0.8.6
python-fity3==0.7
python3-memcached==1.51
PyTweening==1.0.3
pytz==2019.3
pywin32==223
pywinauto==0.6.5
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
retrying==1.3.3
selenium==3.7.0
six==1.12.0
socketIO-client==0.7.2
sqlparse==0.3.0
sty==1.0.0b2
swagger-ui-bundle==0.0.6
text-unidecode==1.3
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
websocket-client==0.56.0
Werkzeug==0.16.0
wrapt==1.11.2
XlsxWriter==1.2.6
yarl==1.4.2
zipp==0.6.0
